(gabac) minor fix missed in #4b93952

+++ src/libs/gabac/stream-handler.cc
@@ -10,7 +10,7 @@
 namespace gabac {
 
 size_t StreamHandler::readStream(std::istream &input, DataBlock *buffer) {
-    uint64_t streamSize = 0;
+    uint32_t streamSize = 0;
     input.read(reinterpret_cast<char *>(&streamSize), sizeof(uint32_t));
     return readBytes(input, streamSize, buffer);
 }
@@ -70,7 +70,7 @@ size_t StreamHandler::readBlock(std::istream &input, size_t bytes, DataBlock *bu
 }
 
 size_t StreamHandler::writeStream(std::ostream &output, DataBlock *buffer) {
-    uint64_t size = buffer->getRawSize();
+    uint32_t size = buffer->getRawSize();
     output.write(reinterpret_cast<char *>(&size), sizeof(uint32_t));
     return writeBytes(output, buffer);
 }
